+++ daml-script/runner/src/main/scala/com/digitalasset/daml/lf/engine/script/v1/Converter.scala
@@ -6,7 +6,7 @@ package engine
 package script
 package v1
 
-import com.daml.ledger.api.v1.transaction.TreeEvent
+import com.daml.ledger.api.v2.transaction.TreeEvent
 import com.daml.ledger.api.v2.transaction.TransactionTree
 import com.digitalasset.canton.ledger.api.validation.NoLoggingValueValidator
 import com.daml.lf.data.Ref._
