+++ daml-script/runner/src/main/scala/com/digitalasset/daml/lf/engine/script/v2/ledgerinteraction/grpcLedgerClient/GrpcLedgerClient.scala
@@ -11,18 +11,18 @@ import org.apache.pekko.stream.Materializer
 import com.daml.grpc.adapter.ExecutionSequencerFactory
 import com.digitalasset.canton.ledger.api.domain.{User, UserRight}
 import com.daml.ledger.api.v2.commands.Commands
-import com.daml.ledger.api.v1.commands._
-import com.daml.ledger.api.v1.event.InterfaceView
+import com.daml.ledger.api.v2.commands._
+import com.daml.ledger.api.v2.event.InterfaceView
 import com.daml.ledger.api.v2.testing.time_service.TimeServiceGrpc.TimeServiceStub
 import com.daml.ledger.api.v2.testing.time_service.{GetTimeRequest, SetTimeRequest, TimeServiceGrpc}
 import com.daml.ledger.api.v2.transaction_filter.TransactionFilter
-import com.daml.ledger.api.v1.transaction_filter.{
+import com.daml.ledger.api.v2.transaction_filter.{
   Filters,
   InclusiveFilters,
   InterfaceFilter,
   TemplateFilter,
 }
-import com.daml.ledger.api.v1.{value => api}
+import com.daml.ledger.api.v2.{value => api}
 import com.daml.lf.CompiledPackages
 import com.digitalasset.canton.ledger.api.validation.NoLoggingValueValidator
 import com.digitalasset.canton.ledger.client.LedgerClient
