+++ language-support/java/bindings-rxjava/src/test/scala/com/daml/ledger/rxjava/grpc/helpers/EventQueryServiceImpl.scala
@@ -3,7 +3,7 @@
 
 package com.daml.ledger.rxjava.grpc.helpers
 
-import com.daml.ledger.api.v1.event_query_service.GetEventsByContractIdRequest
+import com.daml.ledger.api.v2.event_query_service.GetEventsByContractIdRequest
 import com.digitalasset.canton.ledger.api.auth.Authorizer
 import com.digitalasset.canton.ledger.api.auth.services.EventQueryServiceAuthorization
 import com.daml.ledger.api.v2.event_query_service.EventQueryServiceGrpc.EventQueryService
