+++ language-support/java/bindings-rxjava/src/test/scala/com/daml/ledger/rxjava/grpc/helpers/PackageServiceImpl.scala
@@ -3,7 +3,7 @@
 
 package com.daml.ledger.rxjava.grpc.helpers
 
-import com.daml.ledger.api.v1.package_service.{
+import com.daml.ledger.api.v2.package_service.{
   GetPackageResponse,
   GetPackageStatusResponse,
   ListPackagesResponse,
