+++ language-support/java/codegen/src/it/java-1.14/com/daml/GenMapTest.java
@@ -5,7 +5,7 @@ package com.daml;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import com.daml.ledger.api.v1.ValueOuterClass;
+import com.daml.ledger.api.v2.ValueOuterClass;
 import com.daml.ledger.javaapi.data.DamlRecord;
 import com.daml.ledger.javaapi.data.Int64;
 import com.daml.ledger.javaapi.data.Text;
