+++ language-support/java/codegen/src/it/java/com/daml/OptionalTest.java
@@ -5,7 +5,7 @@ package com.daml;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import com.daml.ledger.api.v1.ValueOuterClass;
+import com.daml.ledger.api.v2.ValueOuterClass;
 import com.daml.ledger.javaapi.data.*;
 import com.daml.ledger.javaapi.data.codegen.json.JsonLfDecoder;
 import com.daml.ledger.javaapi.data.codegen.json.JsonLfDecoders;
